+++ genium/src/test/resources/smoke/enums/output/swift/smoke/EnumsInTypeCollection.swift
@@ -2,7 +2,7 @@
 //
 // Automatically generated. Do not modify. Your changes will be lost.
 import Foundation
-public enum TCEnum : UInt32 {
+public enum TCEnum : UInt32, CaseIterable {
     case first
     case second
 }