+++ genium/src/test/resources/smoke/listeners/output/swift/smoke/ListenersWithReturnValues.swift
@@ -128,7 +128,7 @@ internal func copyToCType(_ swiftClass: ListenersWithReturnValues?) -> RefHolder
 internal func moveToCType(_ swiftClass: ListenersWithReturnValues?) -> RefHolder {
     return getRef(swiftClass, owning: true)
 }
-public enum ResultEnum : UInt32 {
+public enum ResultEnum : UInt32, CaseIterable {
     case none
     case result
 }